Consolidate code, remove big endian support
LERC only works on little endian machines anyhow
