+++ gdal/frmts/mrf/GNUmakefile
@@ -41,12 +41,10 @@ XTRA_OPT        :=      $(XTRA_OPT) -I../zlib
 endif
 
 ifeq ($(HAVE_LERC),yes)
-ifneq ("$(wildcard libLERC)","")
 LIBLERC	= libLERC
 OBJ	:=	$(OBJ) LERC_band.o
 XTRA_OPT	:=	$(XTRA_OPT) -DLERC -I$(LIBLERC) -I../../third_party/LercLib
 SUBLIBS := lib-libLERC
-endif
 endif
 
 CPPFLAGS        :=  $(XTRA_OPT)  $(CPPFLAGS)
